Fix incorrect handling of last value index.

+++ core/formats/columnstore2.cpp
@@ -892,7 +892,8 @@ sparse_column::read_blocks_sparse(const column_header& hdr, index_input& in) {
     block.last_size = in.read_long();
     block.last = column::kBlockSize - 1;
   }
-  blocks.back().last = (hdr.docs_count % column::kBlockSize - 1);
+  blocks.back().last =
+      uint16_t(uint32_t(hdr.docs_count % column::kBlockSize - 1));
 
   return blocks;
 }
